dashboard/requirements.txt
celery==3.1.23
cryptography==2.3.1
Django==2.1
django-bootstrap3==10.0.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-registration==2.1.2
djangorestframework==3.8.2
django-cas-ng==3.5.10
gunicorn==19.6.0
jsonpickle==0.9.3
oauth2==1.9.0.post1
oauthlib==1.1.2
pika==0.10.0
psycopg2==2.6.2
PyJWT==1.4.2
requests==2.11.0
django-fernet-fields==0.5
pyyaml==3.13
pytz==2018.5
